Change package location & namespace
