+++ app/src/main/java/org/sopt/official/config/messaging/SoptFirebaseMessagingService.kt
@@ -39,7 +39,7 @@ import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
 import org.sopt.official.R
 import org.sopt.official.network.persistence.SoptDataStore
-import model.UserStatus
+import org.sopt.official.auth.model.UserStatus
 import org.sopt.official.domain.usecase.notification.RegisterPushTokenUseCase
 import org.sopt.official.feature.auth.AuthActivity
 import javax.inject.Inject
