+++ app/src/main/java/org/sopt/official/feature/auth/AuthViewModel.kt
@@ -28,8 +28,8 @@ import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
-import model.Auth
-import model.UserStatus
+import org.sopt.official.auth.model.Auth
+import org.sopt.official.auth.model.UserStatus
 import org.sopt.official.domain.usecase.LoginUseCase
 import timber.log.Timber
 import javax.inject.Inject
