+++ app/src/main/java/org/sopt/official/feature/home/HomeActivity.kt
@@ -47,8 +47,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-import model.UserActiveState
-import model.UserStatus
+import org.sopt.official.auth.model.UserActiveState
+import org.sopt.official.auth.model.UserStatus
 import org.sopt.official.R
 import org.sopt.official.analytics.AmplitudeTracker
 import org.sopt.official.analytics.EventType
