--- core/auth/src/main/kotlin/model/LogOut.kt
@@ -1,7 +1,0 @@
-package model
-
-
-data class LogOut(
-    val success: Boolean,
-    val message: String
-)
