+++ core/auth/src/main/kotlin/org/sopt/official/auth/model/LogOut.kt
@@ -1,0 +1,7 @@
+package org.sopt.official.auth.model
+
+
+data class LogOut(
+    val success: Boolean,
+    val message: String
+)
