--- core/authimpl/src/main/java/org/sopt/official/core/authimpl/api/AuthService.kt
@@ -1,49 +1,0 @@
-/*
- * MIT License
- * Copyright 2022-2023 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.core.authimpl.api
-
-import org.sopt.official.core.authimpl.model.request.AuthRequest
-import org.sopt.official.core.authimpl.model.response.LogOutRequest
-import org.sopt.official.core.authimpl.model.response.LogOutResponse
-import org.sopt.official.network.model.response.AuthResponse
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.HTTP
-import retrofit2.http.POST
-
-interface AuthService {
-    @POST("auth/playground")
-    suspend fun authenticate(
-        @Body body: AuthRequest
-    ): AuthResponse
-
-    @DELETE("user")
-    suspend fun withdraw()
-
-    @HTTP(method = "DELETE", path = "user/logout", hasBody = true)
-    suspend fun logOut(
-        @Body body: LogOutRequest
-    ): LogOutResponse
-}
